Got it to work. But it isn't suuuper good
If the discriminator outputs 1 it means it thinks it is real, 0 is fake. I added a picture of the loss of the models. heuum. make sure to install this if you want to test it, I trained it for a thousand epochs in less than 5 minutes. https://github.com/bstriner/keras-adversarial